+++ docs/changes.rst
@@ -4,6 +4,7 @@ Weblate 4.13.1
 Not yet released.
 
 * Fixed tracking suggestions in history.
+* Fixed parsing reverse proxy info from Cloudflare.
 
 `All changes in detail <https://github.com/WeblateOrg/weblate/milestone/85?closed=1>`__.